Fixed a bug where printing to stdout and stderr would cause a crash because both streams were apparently a string ('nul').
This was caused because when running without a command window, all output is redirected to os.devnull to prevent the stupid windows crash where you cant write to stdout in the absence of the terminal. So I had redirected both streams to os devnul. Little did I know that os.devnull was a string, not a file descriptor, and I needed to open() it to get the file descriptor.