Modified the saving of front panel states to also save the connection table.
We now provide the option to overwrite the front panel state in an existing h5 file, providing the connection tables match. If no connection table is present, we save the existing connection table in.

There is also a silent save feature which we will use at the start of each experiment run which doesn't popup dialog boxes with messages in!