Fixed bugs with NI PCI 6733 and 6363. Turns out if there wasn't any buffered digital output, the clock signal wouldn't be used properly for the analogue output. This was because the static-mode DO had ownership of the PFI which was being used for clocking.
So now when there is no DO, some empty DO is created. We can make this better in the future by having the task not take ownership of the PFI in the first place. In fact I'll do that soon, it's just complicated by the fact that the port/line syntax and PFI refer to the same terminals, so I have to work out which to skip over.