Removed print statements from main.pyw and front_panel_settings.py
Fixed error handling of step size adjustment dialog when an invalid number was given. Added functionality so that hitting enter submits the dialog (equal to clicking OK)

Added right click context menu to spinbutton up/down buttons. This also prevents the default behaviour of the up/down buttons, which is to set the spinbutton to it's upper or lower limit respectively.

For those reading this log, I've also discovered that middle clicking the up/down button adjusts the spinbutton value by the "page size" which is always set to 10 times the step size.