Removed redundant RF output type
Fixed a bug in pulseblaster and novatechdds9m where the channel name was not set to '-' when not in the connection table

Streamlined the pulseblaster __init__ function to iterate over freq/amp/phase when making a DDS object rather than doing each separately!

Added a function to the AO class which takes a settings dictionary and updates the value/properties of the channel to match!