Fixed colour scheme of buttons!
Hover colours now make sense. Colour scheme can be adjusted via the blacs.gtkrc file

Implemented locking for Digital widgets properly. This locking doesn't actually set the widget insensitive, as we need the widget to respond so we can unlock it again! We do a bit of hackery to achieve this, and we need to make sure that we:

A) Don't accidentally program the device to the opposite of the locked value, and program it back (digital edges are bad!)

B) Locked digital widgets update themselves correctly when using the set_state method at the end of the experiment!

Further testing along these lines indicates that the program method gets called even when a widget is locked, but at least it is never asked to program the wrong value (it only ever programs the locked value)

I still need to implement the restoring of the locked state of a widget!