fixes #471
If the self.task object for AI is a valid object within the setup_task() function, we call ClearTask on the object.

This prevents the StartTask call from failing due to a crash on a previous run which failed to cause StopTask to be called.

This bug should now be resolved. Just restart all the crashed tabs (clear BIAS errors if there are any), unpause the queue and you are good to go!!!