Fixed a bug with the new deliberate specification of board_number in the connection table.
If the connection table was defined as:
PulseBlaster('pb_0',board_number=1)
PulseBlaster('pb_1',board_number=0)
then the pulseblasters would be given each others pulse sequences!
Fortunately, this never happened, because we always assigned pb_0 as board 0, etc!

This error was because we picked the group out of the HDF5 file based on the board number (which used to be hard coded and matched to the device name). We now pick the group based on the device name, which is what we should ALWAYS do with every device.