Reversed the order of the 'set units' and 'set value' steps in restoring front panel settings for widgets.
This is because 'set units' doesn't actually happen immediately, it queues up a callback. So by running it first, it puts the object in an inconsistent state where self.units has been updated, but the GUI does not yet reflect this. This causes incorrect results when the set_value method runs, and uses self.units to find out what the units are. It thinks the units have been updated, so it sets the value in the new units.
Then the units actually get updated when the queued up callback runs soon after. This converts the value into the new units -- even though it was already in the new units! The conversion thus results in a nonsese result.

By simply swapping these steps the problem is resolved. The value gets set (in base units), then the units are changed to the non-base units, and the value is converted.