Fixed bug where path name was not updated correctly on shot retry
When a shot fails to run properly, a new HDF5 file is created containing only those components written to the file by runmanager. BLACS attempts to delete the failed shot, and rename the new file to that of the old file. However, on windows sometimes the failed shot file is held open and cannot be deleted. In this case, we append '_retry' to the end of the filename, and add that file to the top of the shot queue. The bug in this case was that the filepath added to the shot queue always corresponded to the old filename, not the new filename