Fixed bug where an experiment with no digital outputs on the NI_pcie_6363 device would not work correctly (analog outputs would not update and it would sometmies raise an exception)
Solution is to make sure you stop the DO task first, to free up the PFI for clock input. For smoe reason it was OK doing this last if you were immediately creating a new DO task, but if you were just stopping the DO task (and not configuring it for a buffered sequence) then it failed.

updated __version__ attribute to be consistent with semver