Queue bugfix: variable 'result' not initialised before used in while loop predicate.
Incidentally a variable with the same name is used earlier (and
I don't believe that code ever sets is to 'done'), which means
this code works most of the time. But that's obivously not ideal.

(I got variable used before assignment when running a shot with no devices -
obviously not a realistic situation but worth fixing anyway)