Functionality to delete repeated shots such that only a finite number of repeated shots are kept on disk at any one time. This prevents consuming disk space in the case of an expeiment being "kept warm" by being left in repeat mode.
This functionality is implemented via a plugin, using a new callback for BLACS
plugins called "shot_complete". The shot deletion is done in a
separate thread so as not to slow down the experiment in the case of blocking
calls, and a zlock lock is acquired on the file prior to deletion
to minimise the harm done if say, a lyse routine is reading the file.

There is sensible behaviour in the case of the user decreasing the number of
shots to retain on disk. Shots that are queued for deletion are kept in a queue,
and when the number of shots to keep is decreased, shots are released from the
front of the queue (without being deleted) until the queue is the
same size as the number of shots now being retained. This means
that if the user set to keep 100 shots, and then goes ahead and runs a
hundred shots, if they then set it to keep 5 shots it won't delete
the 95 oldest shots in the deletion queue. Rather it will only delete the
most recent 5 (and not immediately - over the next 5 shots). I think this is
close to what people will expect when changing the number, and it errs on the
side of keeping shots.