Convert h5 filepaths to network agnostic before sending them to worker processes for transition_to_buffered.
Only do this if the device is actually a remote device. This way, it is backward
compatible with device not expecting remote paths. Workers that are updated to
support being remote workers will need to check self.is_remote before calling
path_to_agnostic, since if running locally the path with continue to be a local one

Bump patch version so that labscript devices can depend on this new behaviour.